web_viewer/backend/requirements.txt
# FastAPI Backend Requirements
# Install with: pip install -r requirements.txt

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=23.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Scientific / medical data
numpy>=1.24.0
pynrrd>=1.0.0
scipy>=1.11.0
trimesh>=4.0.0

# Utilities
python-json-logger>=2.0.0
